Factor out convenience APIs into another class
